EDIT<DTOs>Structure
