+++ ETA.Integrator.Server/Dtos/ConsumerAPI/MetadataParentDTO.cs
@@ -1,0 +1,8 @@
+﻿namespace ETA.Integrator.Server.Dtos.ConsumerAPI
+{
+    public class MetadataParentDTO
+    {
+        public int TotalPages { get; set; }
+        public int TotalCount { get; set; }
+    }
+}
