+++ ETA.Integrator.Server/Dtos/ConsumerAPI/SubmitDocuments/DocumentRejectedDTO.cs
@@ -3,6 +3,6 @@
     public class DocumentRejectedDTO
     {
         public string InternalId { get; set; } = string.Empty;
-        public Error Error { get; set; } = new();
+        public ErrorDTO Error { get; set; } = new();
     }
 }
